Refactor text editor by breaking style buttons and extensions setting into component and file
